White box fix. Huge improvement to control of motion.
Someone on discord informed me that VACE was designed to use white shapes on a black background. Not the other way around. After testing this for about an hour it's clear that white shapes on a black background greatly improves the output.

There is now a huge improvement to the outputs, I will upload comparisons soon.